refactor: group global tests
